Anaface Layout Improvements, Added Environment Dialog, Debug feature for TrecPointers

+++ Anaface/TLayout.h
@@ -397,7 +397,7 @@ protected:
 	/**
 	 * the width/height of the columns/rows
 	 */
-	TDataArray<int> columnLines, rowLines;
+	TDataArray<float> columnLines, rowLines;
 	/**
 	 * Used to determin which columns/rows are flexible
 	 */
