+++ TrecCode/TLanguage.cpp
@@ -386,7 +386,7 @@ bool TLanguage::RunCommentFilter(TrecPointer<TFile> file, TString& newFileName)
 	int endString = -1;
 	DoubIndex sinCom, newL, multiCom, multiComE, sinStr, mulStr;
 
-	while (readData = file->ReadString(inBuffer, (UINT)100))
+	while (readData = file->ReadString(inBuffer, (ULONGLONG)100))
 	{
 		UpdateDoubIndex(inBuffer, 0, sinCom, newL, multiCom, multiComE, sinStr, mulStr);
 		outbuffer.Set(L"");
